Don't cut longer playernames off
Previously names that had more than 12 characters would be cut off. With this change, that does not happen anymore. Also the playernames are still aligned to the right. TODO: Find a proper upper limit for the length of a playername and cut playernames that are longer than that to that length.